Linear Enemy Movement
Added Linear Enemy Movement
Refactored CircleSize To Perfect BeatCircle (Line is when you have to press)
Changed Enemy Movement Formula

+++ Pro5 Game/Assets/Scipts/BeatMachine.cs
@@ -133,7 +133,7 @@ public class BeatMachine : MonoBehaviour
 
     public float getBeatDivided()
     {
-        return beatSec / beatDivider;
+        return (float)beatSec / beatDivider;
     }
 
     public int getBeatDivider()
